+++ lib/active_model/serializer/reflection.rb
@@ -38,6 +38,7 @@ module ActiveModel
         super
         @_links = {}
         @_include_data = true
+        @_load_data = false
         @_meta = nil
       end
 
@@ -56,6 +57,11 @@ module ActiveModel
         :nil
       end
 
+      def load_data(&blk)
+        @_load_data = blk
+        :nil
+      end
+
       # @param serializer [ActiveModel::Serializer]
       # @yield [ActiveModel::Serializer]
       # @return [:nil, associated resource or resource collection]
@@ -69,19 +75,22 @@ module ActiveModel
       #       Blog.find(object.blog_id)
       #     end
       #   end
-      def value(serializer)
+      def value(serializer, current_include_tree)
         @object = serializer.object
         @scope = serializer.scope
 
         if block
-          block_value = instance_exec(serializer, &block)
-          if block_value == :nil
-            serializer.read_attribute_for_serialization(name)
-          else
-            block_value
+          instance_exec(serializer, &block)
+
+          if include_data?(serializer, current_include_tree)
+            if @_load_data
+              @_load_data.call(current_include_tree)
+            else
+              include_data_for(serializer, current_include_tree)
+            end
           end
         else
-          serializer.read_attribute_for_serialization(name)
+          include_data_for(serializer, current_include_tree)
         end
       end
 
@@ -106,11 +115,11 @@ module ActiveModel
       #
       # @api private
       #
-      def build_association(subject, parent_serializer_options)
-        association_value = value(subject)
+      def build_association(subject, parent_serializer_options, current_include_tree = {})
+        association_value = value(subject, current_include_tree)
         reflection_options = options.dup
         serializer_class = subject.class.serializer_for(association_value, reflection_options)
-        reflection_options[:include_data] = @_include_data
+        reflection_options[:include_data] = include_data?(subject, current_include_tree)
 
         if serializer_class
           begin
@@ -133,6 +142,26 @@ module ActiveModel
       attr_accessor :object, :scope
 
       private
+
+      def include_data_for(serializer, current_include_tree)
+        return unless include_data?(serializer, current_include_tree)
+
+        if serializer.class._associations_via_include_param
+          if current_include_tree.key?(name)
+            serializer.read_attribute_for_serialization(name)
+          end
+        else
+          serializer.read_attribute_for_serialization(name)
+        end
+      end
+
+      def include_data?(serializer, current_include_tree)
+        if serializer.class._associations_via_include_param
+          current_include_tree.key?(name)
+        else
+          @_include_data
+        end
+      end
 
       def serializer_options(subject, parent_serializer_options, reflection_options)
         serializer = reflection_options.fetch(:serializer, nil)