+++ test/adapter/json_api/relationships_test.rb
@@ -40,7 +40,7 @@ module ActiveModel
 
             has_many :roles do |serializer|
               meta count: object.posts.count
-              serializer.cached_roles
+              load_data { serializer.cached_roles }
             end
 
             has_one :blog do